Created project
Initial creation of project, with initialisation of project files.
Project will be built upon Phonegap/Cordoba library, with the Ionic
framework on top. Will be created for iOS first and support for Android
and Windows Phone will come later.

Azure mobile service will be used as a backend, and is already set up.